refactors read_writer to save all files at end of pre-processing
Originally, the read_writer would save sensor data and work_table
dataframes before the feature extraction process. This was
intended to be a base data. The PreProcess class would have then
been refactored to allow a user to load this base data rather than
making it from scratch each time.

However, making the base data is quite quick. Plus, saving it at
this point means that the sensor_data and work_table are lacking
information added later in the process. Even though this final
sensor and work_table data is not used after it is saved, it
nonetheless provides users a reference of what occured in the
feature extraction stage