Refactors ML0102Agg Stats to create or import general stats  data
The config_settings for the base factory have been modified to
potentially contain a stats_folder key. The value of this key
is itself a key for the base statistics of that machine in  the
settings dictionary. If the key is present, the PreProcess class
will attempt to load this data. If its not available, the data will
be created and passed to the StatsFeatureExtractor0102Agg
feature_extractor